fix namespace, error msg, move core/filesystem.hpp to io/
